Add anti-stockpiling special case for coal crafting
Added a special case to resourceAvailableForCrafting() for coal, so that it is always has 100% income diversion and 100% target-percentage-crafted.

What this means, in effect, is that 100% of the per-tick coal income will always be treated as being available for crafting, no matter how much steel already exists. However, the coal stockpile level will never _decrease_ unless it was over its normal limit.

This is because coal has no useful purposes except for being crafted into steel, so there's no value in trying to build up a stockpile. However, we don't want to just treat our entire stockpiled coal value as craftable, because that would:
- Make it imposible to control how much of your iron income gets used for steel vs. for plates by adjusting your coal income
- Cause autoTrade to constantly trigger trying to refill your coal stockpile
- Make it dificult to fill your coal stockpile in preparation for performing a reset in order to maximize the amount of coal carried over by chronospheres